job-submission/backend/requirements.txt
fastapi[standard]
python-multipart
motor
python-dotenv
firebase-admin